requirements.txt
requests
python-jenkins
python-gitlab
plotly
yattag
nose